Cleaning up never-reached code.
The java.lang.String#split method never returns an array of zero
elements. If the ID does not contain ':', the array will contain a
single element that is the entire ID. So the `id.Parts.length == 0`
branch can never be reached and is in fact equivalent to the `else`
branch that comes just after.